refactor: create graphql types modify module
