+++ src/blueprint/from_config/graphql/types/mod.rs
@@ -1,7 +1,9 @@
 mod http;
 mod const_field;
 mod unsafe_field;
+mod modify;
 
 pub use http::*;
 pub use const_field::*;
 pub use unsafe_field::*;
+pub use modify::*;
